03-oracle23ai-semantic-search/requirements.txt
openai
oracledb
python-dotenv
numpy